927580: Latest version updated and changed the systemjs to webpack-cli
