+++ .gitignore
@@ -1,3 +1,2 @@
-src/**/*.js
-!src/system.config.js
+dist/
 node_modules/
