--- src/app/conditional-formatting-data.ts
@@ -1,162 +1,0 @@
-export function conditionalFormatData(): Object[] {
-  return [{
-            "Item Code": "AG940Z",
-            "Item Name": "Laser Printer",
-            "Quantity": "144",
-            "Purchase Price": "$169.50",
-            "Selling Price": "$172.00",
-            "Profit": "=E3-D3",
-            "Last Updated": "5/25/19",
-            "Rating": "4.5"
-        },
-        {
-            "Item Code": "BJ120K",
-            "Item Name": "Scientific Calculator",
-            "Quantity": "116",
-            "Purchase Price": "$21.80",
-            "Selling Price": "$23.00",
-            "Profit": "=E4-D4",
-            "Last Updated": "7/28/19",
-            "Rating": "4.0"
-        },
-        {
-            "Item Code": "BC120M",
-            "Item Name": "Wired Keyboard",
-            "Quantity": "438",
-            "Purchase Price": "$26.80",
-            "Selling Price": "$29.00",
-            "Profit": "=E5-D5",
-            "Last Updated": "3/30/20",
-            "Rating": "4.25"
-        },
-        {
-            "Item Code": "BS121L",
-            "Item Name": "Memory Card",
-            "Quantity": "486",
-            "Purchase Price": "$23.50",
-            "Selling Price": "$25.00",
-            "Profit": "=E6-D6",
-            "Last Updated": "8/20/19",
-            "Rating": "3.5"
-        },
-        {
-            "Item Code": "BU121K",
-            "Item Name": "Coffee Maker",
-            "Quantity": "176",
-            "Purchase Price": "$56.50",
-            "Selling Price": "$59.00",
-            "Profit": "=E7-D7",
-            "Last Updated": "2/2/20",
-            "Rating": "4.5"
-        },
-        {
-            "Item Code": "BD121M",
-            "Item Name": "Table Lamp",
-            "Quantity": "0",
-            "Purchase Price": "$22.50",
-            "Selling Price": "$25.00",
-            "Profit": "=E8-D8",
-            "Last Updated": "11/11/19",
-            "Rating": "5.0"
-        },
-        {
-            "Item Code": "AT992X",
-            "Item Name": "Document Scanner",
-            "Quantity": "116",
-            "Purchase Price": "$175.00",
-            "Selling Price": "$177.00",
-            "Profit": "=E9-D9",
-            "Last Updated": "4/13/19",
-            "Rating": "4.75"
-        },
-        {
-            "Item Code": "AP992Z",
-            "Item Name": "Gaming Headset",
-            "Quantity": "58",
-            "Purchase Price": "$32.00",
-            "Selling Price": "$35.00",
-            "Profit": "=E10-D10",
-            "Last Updated": "2/14/20",
-            "Rating": "4.4"
-        },
-        {
-            "Item Code": "AW920X",
-            "Item Name": "Laptop Bag",
-            "Quantity": "232",
-            "Purchase Price": "$18.90",
-            "Selling Price": "$19.00",
-            "Profit": "=E11-D11",
-            "Last Updated": "6/10/19",
-            "Rating": "3.9"
-        },
-        {
-            "Item Code": "AQ920Z",
-            "Item Name": "Table Fan",
-            "Quantity": "405",
-            "Purchase Price": "$33.90",
-            "Selling Price": "$35.00",
-            "Profit": "=E12-D12",
-            "Last Updated": "5/28/19",
-            "Rating": "3.75"
-        },
-        {
-            "Item Code": "AE940X",
-            "Item Name": "Electric Mop",
-            "Quantity": "47",
-            "Purchase Price": "$153.50",
-            "Selling Price": "$155.00",
-            "Profit": "=E13-D13",
-            "Last Updated": "12/18/19",
-            "Rating": "3.9"
-        },
-        {
-            "Item Code": "UI152C",
-            "Item Name": "Smart LED TV",
-            "Quantity": "232",
-            "Purchase Price": "$201.25",
-            "Selling Price": "$204.00",
-            "Profit": "=E14-D14",
-            "Last Updated": "3/13/20",
-            "Rating": "4.35"
-        },
-        {
-            "Item Code": "UD152V",
-            "Item Name": "Robotic Vacuum Cleaner",
-            "Quantity": "210",
-            "Purchase Price": "$182.25",
-            "Selling Price": "$185.00",
-            "Profit": "=E15-D15",
-            "Last Updated": "8/26/19",
-            "Rating": "4.0"
-        },
-        {
-            "Item Code": "BK120L",
-            "Item Name": "Gaming Mouse",
-            "Quantity": "225",
-            "Purchase Price": "$34.80",
-            "Selling Price": "$38.00",
-            "Profit": "=E16-D16",
-            "Last Updated": "1/12/20",
-            "Rating": "4.2"
-        },
-        {
-            "Item Code": "UF162V",
-            "Item Name": "Chair",
-            "Quantity": "373",
-            "Purchase Price": "$104.800",
-            "Selling Price": "$108.00",
-            "Profit": "=E17-D17",
-            "Last Updated": "4/12/19",
-            "Rating": "3.25"
-        },
-        {
-            "Item Code": "UR162C",
-            "Item Name": "Welding Gloves",
-            "Quantity": "216",
-            "Purchase Price": "$19.00",
-            "Selling Price": "$21.00",
-            "Profit": "=E18-D18",
-            "Last Updated": "11/22/19",
-            "Rating": "3.85"
-        }];
-}
