+++ tsconfig.json
@@ -1,22 +1,12 @@
 {
-    "compilerOptions": {
-        "target": "es5",
-        "module": "amd",
-        "removeComments": true,
-        "noLib": false,
-        "sourceMap": true,
-        "pretty": true,
-        "allowUnreachableCode": false,
-        "allowUnusedLabels": false,
-        "noImplicitAny": true,
-        "noImplicitReturns": true,
-        "noImplicitUseStrict": false,
-        "noFallthroughCasesInSwitch": true,
-        "allowJs": false,
-        "noEmitOnError": true,
-        "forceConsistentCasingInFileNames": true,
-        "moduleResolution": "node",
-        "lib": ["es6", "dom"]
-    },
-    "compileOnSave": false
+  "compilerOptions": {
+    "target": "es2016",           
+    "module": "ES6",                                
+    "sourceMap": true,                                
+    "esModuleInterop": true,                             
+    "forceConsistentCasingInFileNames": true,            
+    "strict": true,                                      
+    "skipLibCheck": true,                               
+    "moduleResolution": "node"
+  }
 }
